统一了所有 ID 属性为 Long 类型

+++ app/src/main/java/cn/coegle18/smallsteps/entity/RelationOfBills.kt
@@ -8,8 +8,8 @@ import cn.coegle18.smallsteps.Relation
 @Entity
 data class RelationOfBills(
         val relation: Relation,
-        var mainBill: Int,
-        var relatedBill: Int
+        var mainBill: Long,
+        var relatedBill: Long
 ) {
     @PrimaryKey(autoGenerate = true)
     var relationId: Long = 0
